requirements.txt
streamlit
python-dotenv
requests
langdetect
google-generativeai
langchain
langchain-google-genai
